requirements.txt
transformers
Pillow
torch
